requirements.txt
# TrueRNG Utilities - Python Dependencies
# Pin to minimum compatible versions for stability

pyserial>=3.5
pyusb>=1.3.0
matplotlib>=3.7.0
nltk>=3.8
numpy>=1.24.0
